refactor(benchmark): split the random settings form by concern
random-settings.tsx was one 1537-line component holding four field
groups, 17 form watches and six inline opt-in cards. It is now 234 lines
and 2 watches: one component per concern (SLO targets, load/stages, seed,
distribution, shared prefix, multi-turn), each taking no props and
reading form / action / intl from its own hooks. Editing one card no
longer re-renders the whole form.

SectionCard replaces the bordered card that was inlined six times. `open`
gates both the body and the header gap; passing `onOpenChange` renders
the standard toggle, and a discriminated union keeps that mutually
exclusive with `extra` — four cards put a registered Form.Item there, and
silently dropping one would unregister a field.

Because the gated Form.Items unmount, those cards must derive their open
state with `Form.useWatch(name, { form, preserve: true })`. Plain
useWatch reads getFieldsValue(), which carries REGISTERED fields only, so
it returns undefined while the card is closed — the card could then never
open from prefilled data, and the toggle handler would overwrite it.

Drops dead code found on the way: the sloFieldsFromTargets writes in
writeSloTargets and handleProfileChange (the 9 flat slo_*_ms fields have
no Form.Item, so they never reached the request — handleModalOk expands
slo_targets instead), `currentData` from the form context, and three
unreferenced styles. Raw `display: flex` wrappers became `Flex`.

One behaviour change: the Data Distribution card's expanded state no
longer survives switching the dataset away from Random and back while
all six fields are empty, since the toggle unmounts with the card.
Submitted values are unaffected.
