+++ src/pages/benchmark/forms/dataset.tsx
@@ -1,12 +1,11 @@
 import { PageAction } from '@/config';
 import { AutoTooltip, Select as SealSelect } from '@gpustack/core-ui';
 import { useIntl } from '@umijs/max';
-import { Form, Select, Tag } from 'antd';
+import { Flex, Form, Select, Tag } from 'antd';
 import React from 'react';
 import {
   AUTO_TUNE_DEFAULTS,
   genDatasetSeed,
-  sloFieldsFromTargets,
   sloTargetsFromFields
 } from '../config';
 import { useFormContext } from '../config/form-context';
@@ -20,8 +19,7 @@ import RandomSettingsForm from './random-settings';
 const DatasetForm: React.FC = () => {
   const intl = useIntl();
   const form = Form.useFormInstance();
-  const { action, datasetList, profilesOptions, applyAutoName } =
-    useFormContext();
+  const { action, profilesOptions, applyAutoName } = useFormContext();
   const disabled = action === PageAction.EDIT;
 
   // The blurb under the selector: a strategy tag (Auto-tune for named presets,
@@ -79,10 +77,15 @@ const DatasetForm: React.FC = () => {
       max_seconds: config.max_seconds ?? null,
       request_rate: config.request_rate ?? -1,
       stages: stages,
-      // The 9 flat thresholds are what the API takes; `slo_targets` is the form's
-      // editable view of them and must be re-derived whenever the preset rewrites
-      // them, or the list would still show the previous preset's rows.
-      ...sloFieldsFromTargets(sloTargetsFromFields(config)),
+      // `slo_targets` must be re-derived whenever the preset changes, or the list
+      // would still show the previous preset's rows. Only this field is written:
+      // the 9 flat `slo_*_ms` thresholds the API takes have no Form.Item, so
+      // they could never reach the request (onFinish carries REGISTERED fields
+      // only) — handleModalOk derives them from `slo_targets` instead (see
+      // ../index). Note this is not because store writes are unreachable in
+      // general: getFieldValue, getFieldsValue(true) and a `preserve: true`
+      // useWatch all read the raw store. Nothing here uses any of those on the 9
+      // names, which is what makes the write dead rather than merely unsubmitted.
       slo_targets: sloTargetsFromFields(config)
     });
     applyAutoName?.();
@@ -127,14 +130,7 @@ const DatasetForm: React.FC = () => {
       {/* Strategy tag + description for the picked profile — mirrors the design's
           blurb under the selector so the user knows what the preset does. */}
       {selectedOption && (
-        <div
-          style={{
-            display: 'flex',
-            alignItems: 'flex-start',
-            gap: 8,
-            marginBottom: 20
-          }}
-        >
+        <Flex align="flex-start" gap={8} style={{ marginBottom: 20 }}>
           <Tag
             color={isCustomProfile ? 'green' : 'blue'}
             variant="filled"
@@ -150,16 +146,16 @@ const DatasetForm: React.FC = () => {
                 flex: 1,
                 fontSize: 12,
                 lineHeight: 1.6,
-                color: 'var(--ant-color-text-secondary)'
+                color: 'var(--ant-color-text-tertiary)'
               }}
             >
               {profileDesc}
             </div>
           )}
-        </div>
+        </Flex>
       )}
 
-      <RandomSettingsForm datasetList={datasetList} />
+      <RandomSettingsForm />
     </>
   );
 };
